# docs/openapi.yaml
openapi: 3.0.0
info:
  title: API менеджера аналитики соцсетей (AI)
  description: |
    API для управления источниками в соцсетях, AI‑анализа контента и управления пользователями.
    Содержит эндпоинты аутентификации, управления пользователями, интеграций с соцсетями
    и аналитики на базе LLM.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.yourdomain.com/api/v1
    description: Production server
components:
  securitySchemes:
    OAuth2PasswordBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TopicChainListItem:
      type: object
      properties:
        chain_id:
          type: string
        source_id:
          type: integer
        analyses_count:
          type: integer
        first_date:
          type: string
          format: date
          nullable: true
        last_date:
          type: string
          format: date
          nullable: true
        topics_count:
          type: integer
        topics:
          type: array
          items:
            type: string
        source:
          type: object
          nullable: true
          properties:
            id: { type: integer }
            name: { type: string }
            platform: { type: string }
            platform_type: { type: string }
            external_id: { type: string }
            base_url: { type: string }
...
security:
  - OAuth2PasswordBearer: []
paths:
...
  /dashboard/topic-chains:
    get:
      tags: [dashboard]
      summary: Список цепочек тем
      description: Возвращает цепочки с базовой статистикой и данными источника.
      parameters:
        - in: query
          name: source_id
          schema: { type: integer }
          required: false
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopicChainListItem'
              examples:
                demo:
                  summary: Пример ответа
                  value:
                    - chain_id: "CHAIN_123"
                      source_id: 1
                      analyses_count: 5
                      first_date: "2025-10-01"
                      last_date: "2025-10-10"
                      topics_count: 3
                      topics: ["economy", "inflation", "ruble"]
                      source:
                        id: 1
                        name: "VK Group 1"
                        platform: "VK"
                        platform_type: "vk"
                        external_id: "group1"
                        base_url: "https://vk.com"
...